Fixed flipped coords, outer border support and breakup main.py
